requirements.txt
flask~=2.3.2
flask_cors
SQLAlchemy~=2.0.12
flask_jwt_extended
flask-bcrypt
stripe
python-dotenv
